chore: migrated functions from abgeordnetenwatch.py into dedicated modules
